Implement SIMD-optimized bucket operations and flexible bucket structure for transposition table
- Added `bucket_simd.rs` for SIMD-optimized probe and worst entry finding functions.
- Enhanced `TTEntry` with a priority score calculation method for better replacement decisions.
- Introduced `flexible_bucket.rs` to support variable-sized buckets with optimized probing and storing mechanisms.
- Updated `prefetch.rs` to include adaptive prefetching utilities and statistics tracking.
- Modified `mod.rs` to include new modules and ensure proper usage of the flexible bucket.
- Updated tests to reference the new flexible bucket implementation and validate its functionality.
